requirements.txt
django
django-hstore
django-autocomplete-light
psycopg2
